+++ Documentation/SBV/Examples/TP/InsertionSort.hs
@@ -38,20 +38,20 @@ import qualified Documentation.SBV.Examples.TP.SortHelpers as SH
 
 -- | Insert an element into an already sorted list in the correct place.
 insert :: (OrdSymbolic (SBV a), SymVal a) => SBV a -> SList a -> SList a
-insert = smtFunction "insert" $ \e l -> ite (null l) [e]
+insert = smtRecFunction "insert" (\_ l -> length l) $ \e l -> ite (null l) [e]
                                       $ let (x, xs) = uncons l
                                         in ite (e .<= x) (e .: x .: xs) (x .: insert e xs)
 
 -- | Insertion sort, using 'insert' above to successively insert the elements.
 insertionSort :: (OrdSymbolic (SBV a), SymVal a) => SList a -> SList a
-insertionSort = smtFunction "insertionSort" $ \l -> ite (null l) []
+insertionSort = smtRecFunction "insertionSort" length $ \l -> ite (null l) []
                                                   $ let (x, xs) = uncons l
                                                     in insert x (insertionSort xs)
 
 
 -- | Remove the first occurrence of an number from a list, if any.
 removeFirst :: (Eq a, SymVal a) => SBV a -> SList a -> SList a
-removeFirst = smtFunction "removeFirst" $ \e l -> ite (null l)
+removeFirst = smtRecFunction "removeFirst" (\_ l -> length l) $ \e l -> ite (null l)
                                                       []
                                                       (let (x, xs) = uncons l
                                                        in ite (e .== x) xs (x .: removeFirst e xs))
@@ -59,7 +59,7 @@ removeFirst = smtFunction "removeFirst" $ \e l -> ite (null l)
 -- | Are two lists permutations of each other? Note that we diverge from the counting
 -- based definition of permutation here, since this variant works better with insertion sort.
 isPermutation :: (Eq a, SymVal a) => SList a -> SList a -> SBool
-isPermutation = smtFunction "isPermutation" $ \l r -> ite (null l)
+isPermutation = smtRecFunction "isPermutation" (\l _ -> length l) $ \l r -> ite (null l)
                                                           (null r)
                                                           (let (x, xs) = uncons l
                                                            in x `elem` r .&& isPermutation xs (removeFirst x r))