+++ Documentation/SBV/Examples/TP/SortHelpers.hs
@@ -35,7 +35,7 @@ import Documentation.SBV.Examples.TP.Lists
 
 -- | A predicate testing whether a given list is non-decreasing.
 nonDecreasing :: (OrdSymbolic (SBV a), SymVal a) => SList a -> SBool
-nonDecreasing = smtFunction "nonDecreasing" $ \l ->  null l .|| null (tail l)
+nonDecreasing = smtRecFunction "nonDecreasing" Data.SBV.List.length $ \l ->  null l .|| null (tail l)
                                                  .|| let (x, l') = uncons l
                                                          (y, _)  = uncons l'
                                                      in x .<= y .&& nonDecreasing l'