+++ Documentation/SBV/Examples/TP/StrongInduction.hs
@@ -52,7 +52,7 @@ import Data.SBV.TP
 oddSequence1 :: IO (Proof (Forall "n" Integer -> SBool))
 oddSequence1 = runTP $ do
   let s :: SInteger -> SInteger
-      s = smtFunction "seq" $ \n -> ite (n .<= 0) 1
+      s = smtRecFunction "seq" id $ \n -> ite (n .<= 0) 1
                                   $ ite (n .== 1) 3
                                   $ s (n-2) + 2 * s (n-1)
 
@@ -99,7 +99,7 @@ oddSequence1 = runTP $ do
 oddSequence2 :: IO (Proof (Forall "n" Integer -> SBool))
 oddSequence2 = runTPWith (tpRibbon 50 z3) $ do
   let s :: SInteger -> SInteger
-      s = smtFunction "seq" $ \n -> ite (n .<= 0) 1
+      s = smtRecFunction "seq" id $ \n -> ite (n .<= 0) 1
                                   $ ite (n .== 1) 3
                                   $ 2 * s (n-1) - s (n-2)
 
@@ -149,7 +149,7 @@ oddSequence2 = runTPWith (tpRibbon 50 z3) $ do
 won'tProve1 :: IO ()
 won'tProve1 = runTP $ do
    let len :: SList Integer -> SInteger
-       len = smtFunction "len" $ \xs -> ite (null xs) 0 (1 + len (tail xs))
+       len = smtRecFunction "len" length $ \xs -> ite (null xs) 0 (1 + len (tail xs))
 
    -- Run it for 5 seconds, as otherwise z3 will hang as it can't prove make the inductive step
    _ <- sInductWith z3{extraArgs = ["-t:5000"]} "lengthGood"
@@ -175,7 +175,7 @@ won'tProve1 = runTP $ do
 won'tProve2 :: IO ()
 won'tProve2 = runTP $ do
    let len :: SList Integer -> SInteger
-       len = smtFunction "badLength" $ \xs -> ite (null xs)
+       len = smtRecFunction "badLength" length $ \xs -> ite (null xs)
                                                   123
                                                   (ite (null xs)
                                                        0
@@ -229,7 +229,7 @@ won'tProve4 = runTP $ do
 
    let -- a bizarre (but valid!) way to sum two integers
        weirdSum :: SInteger -> SInteger -> SInteger
-       weirdSum = smtFunction "weirdSum" (\x y -> ite (x .<= 0) y (weirdSum (x - 1) (y + 1)))
+       weirdSum = smtRecFunction "weirdSum" (\x _ -> x) (\x y -> ite (x .<= 0) y (weirdSum (x - 1) (y + 1)))
 
    _ <- sInductWith z3{extraArgs = ["-t:5000"]} "badMeasure"
                 (\(Forall x) (Forall y) -> x .>= 0 .=> weirdSum x y .== x + y)
@@ -277,7 +277,7 @@ sumHalves :: IO (Proof (Forall "xs" [Integer] -> SBool))
 sumHalves = runTP $ do
 
     let halvingSum :: SList Integer -> SInteger
-        halvingSum = smtFunction "halvingSum" $ \xs -> ite (null xs .|| null (tail xs))
+        halvingSum = smtRecFunction "halvingSum" length $ \xs -> ite (null xs .|| null (tail xs))
                                                            (sum xs)
                                                            (let (f, s) = splitAt (length xs `sDiv` 2) xs
                                                             in halvingSum f + halvingSum s)